Move CSV download tests into separate files for readability
